create the .env with the baser before next build
Next picks the .env file at build time and not run time and since we are building our next project in the runner outside the code, it does not have access to the .env before build hence the error. With this, we ensure it has the base url before running build. 

here is a link for more: https://nextjs.org/docs/pages/building-your-application/configuring/environment-variables